rename subfolder, close if broadcasting is finished
