+++ WalletWasabi.Fluent/Views/TransactionBroadcasting/BroadcastTransactionView.axaml
@@ -3,12 +3,12 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
-             xmlns:tB="clr-namespace:WalletWasabi.Fluent.ViewModels.TransactionBroadcaster"
              xmlns:converters="clr-namespace:WalletWasabi.Fluent.Converters"
+             xmlns:transactionBroadcasting="clr-namespace:WalletWasabi.Fluent.ViewModels.TransactionBroadcasting"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:DataType="tB:BroadcastTransactionViewModel"
+             x:DataType="transactionBroadcasting:BroadcastTransactionViewModel"
              x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.TransactionBroadcaster.BroadcastTransactionView">
+             x:Class="WalletWasabi.Fluent.Views.TransactionBroadcasting.BroadcastTransactionView">
     <c:ContentArea Title="Transaction"
                    Caption="{Binding TransactionId}"
                    EnableCancel="True"
