+++ WalletWasabi.Fluent/Views/TransactionBroadcasting/BroadcastTransactionView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.TransactionBroadcaster
+namespace WalletWasabi.Fluent.Views.TransactionBroadcasting
 {
 	public class BroadcastTransactionView : UserControl
 	{
