+++ WalletWasabi.Fluent/Views/TransactionBroadcasting/LoadTransactionView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.TransactionBroadcaster
+namespace WalletWasabi.Fluent.Views.TransactionBroadcasting
 {
 	public class LoadTransactionView : UserControl
 	{
